feat: 修改配置文件和部分代码问题
